+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2018-10-02
+### Changed
+- Rename `ReadDir` method to `Load`.
+- Rename `ReadDirContext` method to `LoadContext`.
+
 ## [0.3.0] - 2018-10-02
 ### Added
 - `ReadDir` and `ReadDirContext` methods for `Cache`.
@@ -26,5 +31,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Source code.
 - Go modules files.
 
+[0.4.0]: https://github.com/gbrlsnchs/filecache/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/gbrlsnchs/filecache/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/gbrlsnchs/filecache/compare/v0.1.0...v0.2.0